refactor : event publisher, handler 객체들을 support.event 디렉토리 하위에 응집 및 event 객체는 domain 계층으로 이동
event는 다른 도메인으로 전파하는 이벤트는 infra 모듈에 속한다고 생각했었다. 하지만, 다른 시스템으로 사건을 전파하는 이벤트와 시스템 내부에서 사용하는 이벤트는 본질적으로 다르다고 생각했고, 다르게 분류해야한다고 생각했다. 이에 따라 시스템 내부에서 발행되는 이벤트는 support.event에 응집하고, 다른 시스템으로 사건을 전파하는 이벤트는 infra.event로 응집해보았다.
추가적으로, event 객체는 도메인 계층에서 상태 변화의 사실을 전달하는 객체라고 생각하여 domain 계층으로 이동시켰다.
