+++ src/main/java/com/pintoss/auth/core/order/application/flow/event/OrderEventPublisher.java
@@ -1,6 +1,6 @@
 package com.pintoss.auth.core.order.application.flow.event;
 
-import com.pintoss.auth.support.event.VoucherPurchaseRequestEvent;
+import com.pintoss.auth.core.order.domain.VoucherPurchaseRequestEvent;
 
 public interface OrderEventPublisher {
     void publish(VoucherPurchaseRequestEvent voucherPurchaseRequestEvent);
