+++ src/main/java/com/pintoss/auth/core/voucher/application/flow/external/VoucherEventPublisher.java
@@ -1,6 +1,6 @@
 package com.pintoss.auth.core.voucher.application.flow.external;
 
-import com.pintoss.auth.support.event.VoucherPurchaseCompletedEvent;
+import com.pintoss.auth.core.voucher.domain.VoucherPurchaseCompletedEvent;
 
 public interface VoucherEventPublisher {
     void publish(VoucherPurchaseCompletedEvent voucherPurchaseCompletedEvent);
