+++ src/main/java/com/pintoss/auth/core/voucher/domain/VoucherPurchaseCompletedEvent.java
@@ -1,4 +1,4 @@
-package com.pintoss.auth.support.event;
+package com.pintoss.auth.core.voucher.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
