+++ src/main/java/com/pintoss/auth/support/event/SpringVoucherEventPublisher.java
@@ -1,6 +1,6 @@
-package com.pintoss.auth.event;
+package com.pintoss.auth.support.event;
 
-import com.pintoss.auth.support.event.VoucherPurchaseCompletedEvent;
+import com.pintoss.auth.core.voucher.domain.VoucherPurchaseCompletedEvent;
 import com.pintoss.auth.core.voucher.application.flow.external.VoucherEventPublisher;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
